Change toughness datatype to Int

+++ simulator/src/main/kotlin/com/primogemstudio/advancedfmk/simulator/objects/interfaces/EnemyObject.kt
@@ -8,7 +8,7 @@ interface EnemyObject {
 
     fun haveWeakness(prop: ObjectWeakness): Boolean
 
-    val initToughness: Short
-    var toughness: Short
+    val initToughness: Int
+    var toughness: Int
     val toughnessBreaked: Boolean
 }
